+++ ndd/utils/__init__.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+# Author: Simone Marsili <simomarsili@gmail.com>
+# License: BSD 3 clause
+"""Utilities."""
+
+from .utils import as_class_name, register_class