+++ CHANGELOG.md
@@ -3,11 +3,10 @@
 ## Unreleased
 
 ### Changed
-- `entropy()` function: renamed `nk` to `counts`
 - removed the `zk` argument
 - the signature of the *entropy* function is:
   ```
-  entropy(counts, k=None, estimator='NSB', return_std=False)
+  entropy(nk, k=None, estimator='NSB', return_std=False)
   ```
 
 ## [1.9.2] - 2020-09-21